Add feature to confirm when deleting article

+++ src/components/Article/DeleteArticle.jsx
@@ -6,17 +6,19 @@ import { deleteObject, ref } from 'firebase/storage';
 
 function DeleteArticle({id, image}) {
     const handleDelete = async() => {
-        try {
-            await deleteDoc(doc(db, "Articles", id))
-            alert("Article Deleted Successfully")
-            
-            //Delete the image from the storage
-            const storageRef = ref(storage, image)
-            await deleteObject(storageRef)
-        } 
-        catch(error) {
-            alert("Error Deleting Article...")
-            console.log(error)
+        if(window.confirm("Are you sure you want to delete this Article?")) {
+            try {
+                await deleteDoc(doc(db, "Articles", id))
+                alert("Article Deleted Successfully")
+                
+                //Delete the image from the storage
+                const storageRef = ref(storage, image)
+                await deleteObject(storageRef)
+            } 
+            catch(error) {
+                alert("Error Deleting Article...")
+                console.log(error)
+            }
         }
     }
